feature(book, guard): create personalizecustomguard
