+++ .env.example
@@ -1,4 +1,4 @@
-PORT=3000
+PORT=3005
 
 DB_NAME=
 DB_USER=
